+++ src/model/verification.rs
