--- src/service/mod.rs
@@ -1,5 +1,0 @@
-pub mod guard;
-pub mod model;
-pub mod pagination;
-pub mod store;
-pub mod time_utils;
