+++ test/factory/idl.d.ts
@@ -1,4 +1,6 @@
 import type { Principal } from '@dfinity/principal';
+import type { ActorMethod } from '@dfinity/agent';
+
 export interface Activity {
   'create_at' : string,
   'canister_id' : Principal,
@@ -116,28 +118,27 @@ export interface Verification {
   'wasm_hash' : [] | [string],
 }
 export interface _SERVICE {
-  'addAdmin' : (arg_0: Principal) => Promise<undefined>,
-  'addBuilder' : (arg_0: Principal) => Promise<undefined>,
-  'addValidator' : (arg_0: Principal) => Promise<undefined>,
-  'deleteAdmin' : (arg_0: Principal) => Promise<undefined>,
-  'deleteBuildConfig' : (arg_0: Principal) => Promise<undefined>,
-  'deleteBuilder' : (arg_0: Principal) => Promise<undefined>,
-  'deleteValidator' : (arg_0: Principal) => Promise<undefined>,
-  'getActivities' : (arg_0: PaginationInfo) => Promise<ManualReply>,
-  'getAdmins' : () => Promise<Array<Principal>>,
-  'getBuildConfigById' : (arg_0: Principal) => Promise<[] | [BuildConfig]>,
-  'getBuildConfigValidator' : (arg_0: BuildConfigInfo) => Promise<
-      [] | [BuildConfig]
-    >,
-  'getBuildConfigs' : () => Promise<Array<BuildConfig>>,
-  'getBuilders' : () => Promise<Array<Principal>>,
-  'getValidators' : () => Promise<Array<Principal>>,
-  'getVerificationByCanisterId' : (arg_0: Principal) => Promise<
-      [] | [Verification]
-    >,
-  'getVerifications' : (arg_0: PaginationInfo) => Promise<ManualReply_1>,
-  'getVerificationsStats' : () => Promise<Stats>,
-  'registerVerification' : (arg_0: RegisterVerification) => Promise<Result>,
-  'saveBuildConfig' : (arg_0: SaveBuildConfig) => Promise<undefined>,
-  'submitVerification' : (arg_0: SubmitVerification) => Promise<undefined>,
+  'addAdmin' : ActorMethod<[Principal], undefined>,
+  'addBuilder' : ActorMethod<[Principal], undefined>,
+  'addValidator' : ActorMethod<[Principal], undefined>,
+  'deleteAdmin' : ActorMethod<[Principal], undefined>,
+  'deleteBuildConfig' : ActorMethod<[Principal], undefined>,
+  'deleteBuilder' : ActorMethod<[Principal], undefined>,
+  'deleteValidator' : ActorMethod<[Principal], undefined>,
+  'getActivities' : ActorMethod<[PaginationInfo], ManualReply>,
+  'getAdmins' : ActorMethod<[], Array<Principal>>,
+  'getBuildConfigById' : ActorMethod<[Principal], [] | [BuildConfig]>,
+  'getBuildConfigValidator' : ActorMethod<
+    [BuildConfigInfo],
+    [] | [BuildConfig],
+  >,
+  'getBuildConfigs' : ActorMethod<[], Array<BuildConfig>>,
+  'getBuilders' : ActorMethod<[], Array<Principal>>,
+  'getValidators' : ActorMethod<[], Array<Principal>>,
+  'getVerificationByCanisterId' : ActorMethod<[Principal], [] | [Verification]>,
+  'getVerifications' : ActorMethod<[PaginationInfo], ManualReply_1>,
+  'getVerificationsStats' : ActorMethod<[], Stats>,
+  'registerVerification' : ActorMethod<[RegisterVerification], Result>,
+  'saveBuildConfig' : ActorMethod<[SaveBuildConfig], undefined>,
+  'submitVerification' : ActorMethod<[SubmitVerification], undefined>,
 }
